np.isnan is now acceptable for isSignifantlylower + beauty change in AS plot

+++ Code/developmental_functions.py
@@ -660,6 +660,9 @@ def simplePTUmerge(infolder, outfolder = ''):
             if file.endswith('.ptu'):
                 source = os.path.join(subfolder, file)
                 destination = os.path.join(outfolder, file)
+                # this avoids copying files twice
+                if os.path.isfile(destination): 
+                    continue
                 shutil.copyfile(source, destination)
 
 def filterFRETind(locLst, indicator, vmin, vmax):
